common: move huge_page_pool to a dedicated header.
